chore: Centralize the snapshot tests for arrays under core (#3535)
* chore: Centralize the snapshot tests for arrays under core
To minimize the duplication of snapshot testing code across themes, centralized it in core in a manner similar to how schema tests work with utils and the validators

* Update packages/docs/docs/contributing.md


---------
