+++ packages/core/test/ArraySnap.test.jsx
@@ -1,0 +1,4 @@
+import Form from '../src';
+import arrayTests from '../testSnap/arrayTests';
+
+arrayTests(Form);
