+++ packages/semantic-ui/test/Array.test.tsx
@@ -1,6 +1,4 @@
-import { ErrorSchema, RJSFSchema, UiSchema } from '@rjsf/utils';
-import validator from '@rjsf/validator-ajv8';
-import renderer from 'react-test-renderer';
+import arrayTests from '@rjsf/core/testSnap/arrayTests';
 
 import Form from '../src';
 
@@ -10,85 +8,4 @@ jest.mock('@fluentui/react-component-ref', () => ({
   Ref: jest.fn().mockImplementation(({ children }) => children),
 }));
 
-describe('array fields', () => {
-  test('array', () => {
-    const schema: RJSFSchema = {
-      type: 'array',
-      items: {
-        type: 'string',
-      },
-    };
-    const tree = renderer.create(<Form schema={schema} validator={validator} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-  test('fixed array', () => {
-    const schema: RJSFSchema = {
-      type: 'array',
-      items: [
-        {
-          type: 'string',
-        },
-        {
-          type: 'number',
-        },
-      ],
-    };
-    const tree = renderer.create(<Form schema={schema} validator={validator} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-  test('checkboxes', () => {
-    const schema: RJSFSchema = {
-      type: 'array',
-      items: {
-        type: 'string',
-        enum: ['a', 'b', 'c'],
-      },
-      uniqueItems: true,
-    };
-    const tree = renderer.create(<Form schema={schema} validator={validator} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-  test('array icons', () => {
-    const schema: RJSFSchema = {
-      type: 'array',
-      items: {
-        type: 'string',
-      },
-    };
-    const uiSchema: UiSchema = {
-      'ui:options': { copyable: true },
-    };
-    const tree = renderer
-      .create(<Form schema={schema} uiSchema={uiSchema} validator={validator} formData={['a', 'b']} />)
-      .toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-  test('no errors', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        name: {
-          type: 'string',
-        },
-      },
-    };
-    const tree = renderer.create(<Form schema={schema} validator={validator} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-  test('empty errors array', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        name: {
-          type: 'string',
-        },
-      },
-    };
-    const errors: any[] = [];
-    const extraErrors = {
-      name: { __errors: errors },
-    } as unknown as ErrorSchema;
-    const tree = renderer.create(<Form schema={schema} validator={validator} extraErrors={extraErrors} />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
+arrayTests(Form);
